+++ core/LFO.h
@@ -14,10 +14,10 @@ class LFO: public ILFO {
 public:
     LFO();
     
-    double amplitude() override { return _amp; }
-    void setAmplitude(double) override;
-    void setPeriod(double) override;
-    void setCurrentBeat(double) override;
+    float amplitude() override { return _amp; }
+    void setAmplitude(float) override;
+    void setPeriod(float) override;
+    void setCurrentBeat(float) override;
     void setFramesPerMeasure(long) override;
     float triangleValueAt(int) override;
     
@@ -25,9 +25,9 @@ private:
     bool _isOn { false };
     long _framesPerMeasure { 0 };
     long _framesPerBeat { 0 };
-    double _currentBeat { 0 };
-    double _period { 0.25 }; // 1/4
-    double _amp { 1.0 }; // +1.0...-1.0
+    float _currentBeat { 0 };
+    float _period { 0.25 }; // 1/4
+    float _amp { 1.0 }; // +1.0...-1.0
  };
 
 #endif /* LFO_h */
